+++ src/client/src/api.ts
@@ -6,6 +6,8 @@ axios.defaults.headers = {
     },
 };
 
+export const serverPath = 'http://localhost:5000/';
+
 export default axios.create({
     baseURL: 'http://localhost:5000/api/',
     withCredentials: true,
